HBX-3068: Refactor the Ant integration tests to factor out common code   - Pull up methods 'getConsoleLogger()' and 'runAntBuild()'   - Get rid of the 'output' ByteArrayStream
